Delay fix
I discovered that delay() functions don't work inside the Ticker since it's functions are using the underlying timer.
Therefore I wrote a custom delay function that is just stuck inside a while loop untill the delay time has passed. This works, but now the Ticker times can become a bit inconsistent. To fix this a timeUpdateReal() method was introduced that updates with real time passed instead of a constant variable.