Stopping conditions + Pulsating charge
Stopping conditions were added to automatically stop the charger. The fan turns on when certain temperature is met and stays on until temperature is decreased to a specific temperature.

If the battery is nearly completely empty, the voltage difference between charger input and battery is very high, because of which the voltage regulator has to do alot of work to keep the charger at supplying exactly 1A. Therefore the regulator could get too hot and the fan isn't able to keep up.
To make sure charging still continues pulsating charge was added. When the temperature reaches a certain treshold the charger is turned of for a small part of a cycle. The higher the temperature the longer the charger is turned off (Similar to PWM).