Fan boost
The emptier the battery is, the hotter the charger will become. When connecting an empty battery, turning the fan on and off is not beneficial as the time the fan will remain off is very small (a few seconds), which allows the charger to become hotter. To prevent this I've introduced a fan boost mode. The first 10 minutes the charger is turned on the charger will consistently remain on..